making oodb
