+++ database/db_answer.py
@@ -1,0 +1,17 @@
+from enum import IntEnum
+
+class DBAnswer:
+    def __init__(self, ok, error, info):
+        self.ok = ok
+        self.error = error
+        self.info = info
+
+
+class DBError(IntEnum):
+    AlwaysOk = 0
+    IncorrectUsername = 10
+    IncorrectPassword = 11
+    IncorrectAvatar = 12
+    NoSuchUser = 20
+
+OK = DBError.AlwaysOk
